Move modifier/keyword related funcs into a separate file
